Swap panda/skeleton skin id, Clean up config file, move stupid ones to config.extra.ini, Moved script registration calls to a separate file, Removed scriptdictionary addon (possibly temporarily), true/yes values for booleans in config files (variant::operator bool), Remove comma/semicolon prefix requirement on serialized lists, Precompiled header files for common includes and forward declare files, Two extra columns for characters for script and quest variable storage
git-svn-id: svn://svn.eoserv.net/eoserv/trunk@168 da1c5d5e-72c9-4536-8af9-af1c9134b0e6

+++ src/eoserver.cpp
@@ -6,20 +6,14 @@
 
 #include "eoserver.hpp"
 
-#include <string>
-
-#include <pthread.h>
-
-#include "world.hpp"
-
-#include "socket.hpp"
-#include "packet.hpp"
-#include "eoclient.hpp"
-#include "timer.hpp"
-#include "nanohttp.hpp"
-#include "util.hpp"
-#include "sln.hpp"
 #include "console.hpp"
+#include "eoclient.hpp"
+#include "nanohttp.hpp"
+#include "packet.hpp"
+#include "sln.hpp"
+#include "socket.hpp"
+#include "timer.hpp"
+#include "world.hpp"
 
 #define CLIENT_F_HANDLE(ID,FUNC) \
 case ID: \
@@ -137,7 +131,7 @@ void EOServer::Initialize(util::array<std::string, 5> dbinfo, const Config &eose
 
 	this->world->server = this;
 
-	if (static_cast<int>(this->world->config["SLN"]))
+	if (this->world->config["SLN"])
 	{
 		this->sln = new SLN(this);
 	}
