+++ src/handlers/Welcome.cpp
@@ -4,7 +4,12 @@
  * See LICENSE.txt for more info.
  */
 
-#include "handlers.hpp"
+#include "handlers.h"
+
+#include "eodata.hpp"
+#include "guild.hpp"
+#include "map.hpp"
+#include "npc.hpp"
 
 CLIENT_F_FUNC(Welcome)
 {
@@ -50,7 +55,7 @@ CLIENT_F_FUNC(Welcome)
 			reply.AddInt(this->player->character->id);
 			reply.AddShort(this->player->character->mapid); // Map ID
 
-			if (static_cast<int>(this->server->world->config["GlobalPK"]) && !this->server->world->PKExcept(this->player->character->mapid))
+			if (this->server->world->config["GlobalPK"] && !this->server->world->PKExcept(this->player->character->mapid))
 			{
 				reply.AddByte(0xFF);
 				reply.AddByte(0x01);
@@ -361,7 +366,7 @@ CLIENT_F_FUNC(Welcome)
 				char buf[4096];
 				int len = std::fread(buf, sizeof(char), 4096, fh);
 
-				if (file == FILE_MAP && static_cast<int>(this->server->world->config["GlobalPK"]) && !this->server->world->PKExcept(this->player->character->mapid))
+				if (file == FILE_MAP && this->server->world->config["GlobalPK"] && !this->server->world->PKExcept(this->player->character->mapid))
 				{
 					if (p + len >= 0x04 && 0x03 - p > 0) buf[0x03 - p] = 0xFF;
 					if (p + len >= 0x05 && 0x04 - p > 0) buf[0x04 - p] = 0x01;
@@ -383,7 +388,7 @@ CLIENT_F_FUNC(Welcome)
 			reply.AddString(content);
 			CLIENT_SENDRAW(reply);
 
-			if (static_cast<int>(this->server->world->config["ProtectMaps"]))
+			if (this->server->world->config["ProtectMaps"])
 			{
 				reply.Reset();
 				reply.AddChar(INIT_BANNED);
